Move some code around.
