+++ tests/phpunit/tests/Shortcode/Embed_Shortcode.php
@@ -15,16 +15,14 @@
  * limitations under the License.
  */
 
-namespace Google\Web_Stories\Tests\Block;
-
-use WP_Block_Type_Registry;
+namespace Google\Web_Stories\Tests\Shortcode;
 
 /**
- * @coversDefaultClass \Google\Web_Stories\Block\Embed_Block
+ * @coversDefaultClass \Google\Web_Stories\Shortcode\Embed_Shortcode
  */
-class Embed_Block extends \WP_UnitTestCase {
+class Embed_Shortcode extends \WP_UnitTestCase {
 	public function tearDown() {
-		unregister_block_type( \Google\Web_Stories\Block\Embed_Block::BLOCK_NAME );
+		remove_shortcode( \Google\Web_Stories\Shortcode\Embed_Shortcode::SHORTCODE_NAME );
 
 		parent::tearDown();
 	}
@@ -33,18 +31,18 @@ class Embed_Block extends \WP_UnitTestCase {
 	 * @covers ::init
 	 */
 	public function test_registers_block_type() {
-		$this->assertTrue( WP_Block_Type_Registry::get_instance()->is_registered( \Google\Web_Stories\Block\Embed_Block::BLOCK_NAME ) );
+		$this->assertTrue( shortcode_exists( \Google\Web_Stories\Shortcode\Embed_Shortcode::SHORTCODE_NAME ) );
 	}
 
 	/**
-	 * @covers ::render_block
+	 * @covers ::render_shortcode
 	 * @covers \Google\Web_Stories\Embed_Base::render
 	 * @covers \Google\Web_Stories\Story_Renderer\Embed::render
 	 */
-	public function test_render_block() {
-		$embed_block = new \Google\Web_Stories\Block\Embed_Block();
+	public function test_render_shortcode() {
+		$embed_shortcode = new \Google\Web_Stories\Shortcode\Embed_Shortcode();
 
-		$actual = $embed_block->render_block(
+		$actual = $embed_shortcode->render_shortcode(
 			[
 				'url'    => 'https://example.com/story.html',
 				'title'  => 'Example Story',
@@ -59,14 +57,14 @@ class Embed_Block extends \WP_UnitTestCase {
 	}
 
 	/**
-	 * @covers ::render_block
+	 * @covers ::render_shortcode
 	 * @covers \Google\Web_Stories\Embed_Base::render
 	 * @covers \Google\Web_Stories\Story_Renderer\Embed::render
 	 */
-	public function test_render_block_missing_url() {
-		$embed_block = new \Google\Web_Stories\Block\Embed_Block();
+	public function test_render_shortcode_missing_url() {
+		$embed_shortcode = new \Google\Web_Stories\Shortcode\Embed_Shortcode();
 
-		$actual = $embed_block->render_block(
+		$actual = $embed_shortcode->render_shortcode(
 			[
 				'url'    => '',
 				'title'  => 'Example Story',
@@ -81,14 +79,14 @@ class Embed_Block extends \WP_UnitTestCase {
 	}
 
 	/**
-	 * @covers ::render_block
+	 * @covers ::render_shortcode
 	 * @covers \Google\Web_Stories\Embed_Base::render
 	 * @covers \Google\Web_Stories\Story_Renderer\Embed::render
 	 */
-	public function test_render_block_missing_title() {
-		$embed_block = new \Google\Web_Stories\Block\Embed_Block();
+	public function test_render_shortcode_missing_title() {
+		$embed_shortcode = new \Google\Web_Stories\Shortcode\Embed_Shortcode();
 
-		$actual = $embed_block->render_block(
+		$actual = $embed_shortcode->render_shortcode(
 			[
 				'url'    => 'https://example.com/story.html',
 				'title'  => '',
@@ -103,16 +101,16 @@ class Embed_Block extends \WP_UnitTestCase {
 	}
 
 	/**
-	 * @covers ::render_block
+	 * @covers ::render_shortcode
 	 * @covers \Google\Web_Stories\Embed_Base::render
 	 * @covers \Google\Web_Stories\Story_Renderer\Image::render
 	 */
-	public function test_render_block_feed_no_poster() {
-		$embed_block = new \Google\Web_Stories\Block\Embed_Block();
+	public function test_render_shortcode_feed_no_poster() {
+		$embed_shortcode = new \Google\Web_Stories\Shortcode\Embed_Shortcode();
 
 		$this->go_to( '/?feed=rss2' );
 
-		$actual = $embed_block->render_block(
+		$actual = $embed_shortcode->render_shortcode(
 			[
 				'url'   => 'https://example.com/story.html',
 				'title' => 'Example Story',
@@ -125,17 +123,17 @@ class Embed_Block extends \WP_UnitTestCase {
 	}
 
 	/**
-	 * @covers ::render_block
+	 * @covers ::render_shortcode
 	 * @covers \Google\Web_Stories\Embed_Base::render
 	 * @covers \Google\Web_Stories\Story_Renderer\Image::render
 	 */
-	public function test_render_block_with_poster() {
-		$embed_block = new \Google\Web_Stories\Block\Embed_Block();
-		$embed_block->init();
+	public function test_render_shortcode_with_poster() {
+		$embed_shortcode = new \Google\Web_Stories\Shortcode\Embed_Shortcode();
+		$embed_shortcode->init();
 
 		$this->go_to( '/?feed=rss2' );
 
-		$actual = $embed_block->render_block(
+		$actual = $embed_shortcode->render_shortcode(
 			[
 				'url'    => 'https://example.com/story.html',
 				'title'  => 'Example Story',
